+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/DiscordBot.kt
@@ -76,10 +76,10 @@ internal class DiscordBotImpl(
     }
 
     private suspend fun startKord() {
-//        cleanOldGuildCommands(kord)
-//        createGlobalCommands()
+        cleanOldGuildCommands(kord)
+        createGlobalCommands()
 //        createAdminCommands()
-        createCommandsForTestServer()
+//        createCommandsForTestServer()
 
         monitorGatewayHealth()