+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/DiscordRegisteredFeature.kt
@@ -21,7 +21,7 @@ internal interface DiscordRegisteredFeature {
 
     suspend fun execute(
         command: Command,
-        query: String,
+        query: String?,
         origin: Source,
     ): Result<BotOutput, BotError>
 }