+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/FeatureRegistry.kt
@@ -2,14 +2,13 @@ package io.github.sophon.discord.feat
 
 internal class FeatureRegistry(
     features: List<DiscordRegisteredFeature>,
-//    private val coreFeature: DiscordRegisteredFeature,
+    private val coreFeature: DiscordRegisteredFeature,
 ) {
     private val featureMap = features.associateBy { it.featureInfo.name }
 
     fun getFeatures(names: List<String>): List<DiscordRegisteredFeature> {
         val configuredFeatures = names.mapNotNull { getFeature(it) }
-//        return configuredFeatures + coreFeature
-        return configuredFeatures
+        return listOf(coreFeature) + configuredFeatures
     }
 
     fun getRegisteredFeatures(): List<DiscordRegisteredFeature> {