+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/emoji/Emojifier.kt
@@ -1,8 +1,7 @@
 package io.github.sophon.discord.feat.emoji
 
 import io.github.aakira.napier.Napier
-import io.github.sophon.core.domain.model.Property
-import io.github.sophon.core.domain.model.ScaledValue
+import io.github.sophon.core.domain.model.Effect
 import io.github.sophon.discord.feat.emoji.data.DiscordEmojiSource
 import io.github.sophon.discord.feat.emoji.data.toFormatted
 
@@ -18,10 +17,50 @@ internal class Emojifier(
         Napier.d(tag = TAG) { "Emojis: ${emojiMap.keys.size}" }
     }
 
-    fun emojify(property: Property): String {
-        //TODO: implement String.emojify(): Emoji
+    fun emojify(effect: Effect): String {
+        //TODO: purely relying on Type won't work, we need to combine both
+        val emoji = when (effect.type) {
+            "fire_rate" -> Emoji.BULLET_SPEED
+            "health" -> Emoji.HP
+            "healing" -> Emoji.HEAL_LIFESTEAL
+            "clipsize" -> Emoji.CLIP
+            "duration" -> Emoji.COOLDOWN_DELAY_DURATION
+            "cooldown",
+            "charge_cooldown",
+                -> Emoji.ABILITY_CD
 
-        return property.type ?: ""
+            "sprint" -> Emoji.SPRINT_SPEED
+            "move_speed" -> Emoji.MOVE_SPEED
+
+            "cast" -> Emoji.CHANNEL_TIME
+
+            "tech_armor_up" -> Emoji.SPIRIT_RESIST
+            "bullet_armor_up",
+            "bullet_armor_down",
+                -> Emoji.BULLET_RESIST
+            "damage" -> Emoji.DAMAGE_AMPLIFY
+
+            "tech_damage" -> Emoji.SPIRIT_DAMAGE
+            "melee_damage" -> Emoji.MELEE_DAMAGE
+            "bullet_damage" -> Emoji.BULLET_WEAPON_DAMAGE
+
+            "distance",
+            "range",
+                -> Emoji.RANGE
+
+            "slow" -> Emoji.MOVEMENT_SLOW
+
+            else -> {
+                if (effect.type.isNullOrBlank().not()) {
+                    Napier.i(tag = TAG) { "unhandled: ${effect.type}" }
+                }
+                null
+            }
+        }
+
+        val result = emoji?.let { emojify(it) } ?: ""
+
+        return result
     }
 
     fun emojify(emoji: Emoji): String {