+++ feat/deadlock/impl/src/commonMain/kotlin/io/github/sophon/deadlock/deadlockModule.kt
@@ -9,10 +9,11 @@ import io.github.sophon.deadlock.db.ItemDatabaseImpl
 import io.github.sophon.deadlock.usecase.SyncAbilitiesUseCase
 import io.github.sophon.deadlock.usecase.SyncDataUseCase
 import io.github.sophon.deadlock.usecase.SyncHeroesUseCase
-import io.github.sophon.deadlock.usecase.SyncItemsUseCase
+import io.github.sophon.deadlock.usecase.DownloadItemsUseCase
 import io.github.sophon.deadlock.remote.DeadlockWikiDataSource
 import io.github.sophon.deadlock.remote.DeadlockWikiDataSourceImpl
 import io.github.sophon.deadlock.remote.ImageResolver
+import io.github.sophon.deadlock.remote.ImageResolverImpl
 import io.github.sophon.deadlock.usecase.DownloadDescriptionsUseCase
 import io.github.sophon.deadlock.usecase.FetchAbilityUseCase
 import io.github.sophon.deadlock.usecase.FetchHeroUseCase
@@ -30,11 +31,11 @@ fun deadlockModule() = module {
     singleOf(::AbilityDatabaseImpl).bind<AbilityDatabase>()
     singleOf(::HeroDatabaseImpl).bind<HeroDatabase>()
     singleOf(::ItemDatabaseImpl).bind<ItemDatabase>()
-    singleOf(::ImageResolver)
+    singleOf(::ImageResolverImpl).bind<ImageResolver>()
 
     singleOf(::SyncAbilitiesUseCase)
     singleOf(::SyncHeroesUseCase)
-    singleOf(::SyncItemsUseCase)
+    singleOf(::DownloadItemsUseCase)
     singleOf(::SyncDataUseCase)
     singleOf(::FetchItemUseCase)
     singleOf(::FetchHeroUseCase)