+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [v0.3.0] - 2026-04-23
+- upgrade path for items
+- emojified Ability and Item properties
+- enabled slash commands
+- `invite`, `help` and `repo` commands
+
 ## [v0.2.0] - 2026-04-19
 - rework Item and Ability data structures
   - better data and with descriptions